requirements.txt
requests 
opencv-python 
Pillow 
pyttsx3
streamlit